[REFACTOR] 도메인형에 맞추어 패키지 구조 수정
