+++ nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.nonsoolmate.nonsoolmateServer.common.swagger;
+package com.nonsoolmate.nonsoolmateServer.global.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Info;
